+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Renamed `Create/EditMessageOptions.Embed` to `Embeds` to support multiple embeds permessage.
 - `DiscordHttpClient.Create/EditMessage` file attachment overloads have been removed. Instead, please use the new `Attachments` property of `Create/EditMessageOptions`.
 - Removed `DiscordHttpClient.UseSingleHttpClient` (it's no longer needed).
+- Removed deprecated `DiscordHttpClient.GetUserDMs`.
 
 ### Additions
 - Added support for creating and editing messages with multiple embeds and/or multiple attachments.